requirements.txt
beautifulsoup4==4.12.3
chromadb==0.6.3
google-genai==1.5.0
jinja2==3.1.4
krippendorff==0.8.0
matplotlib==3.10.6
newspaper3k==0.2.8
numpy==1.26.4
openai==1.97.1
pandas==2.3.2
pydantic==2.9.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
tenacity==9.0.0
tldextract==5.3.0
together==1.4.6
typing_extensions==4.12.2
